Starting permissions feature.
Starting to work on the custom permissions feature. This will add a third permission type in addition to public and private. The `custom` permission type will be used for configurable commands, this will be useful for user role assignments and tag creation so far. Banning, muting and kicking will benefit from this as well.